+++ pdf/src/main/groovy/com/craigburke/document/builder/render/TableRenderer.groovy
@@ -37,7 +37,7 @@ class TableRenderer implements Renderable {
         }
 
         boolean reachedEnd = false
-        float remainingHeight = height - (onFirstPage ? table.border.size : 0)
+        float remainingHeight = (height - (onFirstPage ? table.border.size : 0)).floatValue()
 
         while (!reachedEnd) {
             RowRenderer currentRenderer = rowRenderers[parseEnd]
